radius modification and dx, dy set to global
